requirements.txt
openai-agents>=0.6.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
